third_party/blink/web_tests/platform/generic/http/tests/devtools/sources/debugger-frameworks/frameworks-dom-xhr-event-breakpoints-expected.txt

Tests framework black-boxing on DOM, XHR and Event breakpoints.


Running: testDOMBreakpoint
Call stack:
  * 0) Framework.appendChild (framework.js:72)
    1) appendElement (frameworks-dom-xhr-event-breakpoints.js:22)
    2)  (:1)
    [setTimeout]
    0)  (:1)
Paused on subtree modificationsdiv#rootElementChild div#rootElement added

Running: testXHRBreakpoint
Call stack:
  * 0) Framework.sendXHR (framework.js:79)
    1) sendXHR (frameworks-dom-xhr-event-breakpoints.js:27)
    2)  (:1)
    [setTimeout]
    0)  (:1)
Paused on XHR or fetchhttp://127.0.0.1:8000/devtools/sources/debugger-frameworks/resources/foo?xhr

Running: testEventListenerBreakpoint
Call stack:
    0) testElementClicked (frameworks-dom-xhr-event-breakpoints.js:34)
  * 1) Framework_bound (framework.js:105)
  * 2) Framework_eventListener (framework.js:86)
    3) addListenerAndClick (frameworks-dom-xhr-event-breakpoints.js:45)
    4)  (:1)
    [setTimeout]
    0)  (:1)

Executing Resume...
Call stack:
    0) testElementClicked (frameworks-dom-xhr-event-breakpoints.js:34)
  * 1) Framework_bound (framework.js:105)
  * 2) Framework_eventListener (framework.js:86)
    3) addListenerAndClick (frameworks-dom-xhr-event-breakpoints.js:51)
    4)  (:1)
    [setTimeout]
    0)  (:1)

Executing Resume...

Running: testSteppingThroughEventListenerBreakpoint
Executing StepOver...
Call stack:
    0) addListenerAndClick (frameworks-dom-xhr-event-breakpoints.js:41)
    1)  (:1)
    [setTimeout]
    0)  (:1)

Executing StepOver...
Call stack:
    0) addListenerAndClick (frameworks-dom-xhr-event-breakpoints.js:42)
    1)  (:1)
    [setTimeout]
    0)  (:1)

Executing StepOver...
Executing StepOver...
Executing StepOver...
Call stack:
    0) testElementClicked (frameworks-dom-xhr-event-breakpoints.js:34)
  * 1) Framework_bound (framework.js:105)
  * 2) Framework_eventListener (framework.js:86)
    3) addListenerAndClick (frameworks-dom-xhr-event-breakpoints.js:45)
    4)  (:1)
    [setTimeout]
    0)  (:1)

Executing StepOut...
Executing StepOver...
Executing StepOver...
Executing StepOver...
Executing StepOver...
Call stack:
    0) testElementClicked (frameworks-dom-xhr-event-breakpoints.js:34)
  * 1) Framework_bound (framework.js:105)
  * 2) Framework_eventListener (framework.js:86)
    3) addListenerAndClick (frameworks-dom-xhr-event-breakpoints.js:51)
    4)  (:1)
    [setTimeout]
    0)  (:1)

Executing StepOver...
Executing StepOver...
Executing StepOver...
Call stack:
    0) addListenerAndClick (frameworks-dom-xhr-event-breakpoints.js:53)
    1)  (:1)
    [setTimeout]
    0)  (:1)

Executing Resume...

Running: testSteppingOutOnEventListenerBreakpoint
Executing StepOut...
Call stack:
    0) testElementClicked (frameworks-dom-xhr-event-breakpoints.js:34)
  * 1) Framework_bound (framework.js:105)
  * 2) Framework_eventListener (framework.js:86)
    3) addListenerAndClick (frameworks-dom-xhr-event-breakpoints.js:45)
    4)  (:1)
    [setTimeout]
    0)  (:1)

Executing StepOut...
Executing StepOut...
Call stack:
    0) testElementClicked (frameworks-dom-xhr-event-breakpoints.js:34)
  * 1) Framework_bound (framework.js:105)
  * 2) Framework_eventListener (framework.js:86)
    3) addListenerAndClick (frameworks-dom-xhr-event-breakpoints.js:51)
    4)  (:1)
    [setTimeout]
    0)  (:1)

Executing StepOut...
Call stack:
    0) addListenerAndClick (frameworks-dom-xhr-event-breakpoints.js:52)
    1)  (:1)
    [setTimeout]
    0)  (:1)

Executing Resume...

Running: testSteppingOutOnEventListenerBreakpointAllBlackboxedButOne
Executing StepOut...
Call stack:
    0) testElementClicked (frameworks-dom-xhr-event-breakpoints.js:60)
  * 1) Framework.safeRun (framework.js:8)
  * 2) Framework.safeRun (framework.js:10)
  * 3) Framework_bound (framework.js:105)
  * 4) Framework_eventListener (framework.js:86)
    5) inner (frameworks-dom-xhr-event-breakpoints.js:72)
    6) addFewBlackboxedListenersAndClick (frameworks-dom-xhr-event-breakpoints.js:78)
    7)  (:1)
    [setTimeout]
    0)  (:1)

Executing StepOut...
Call stack:
    0) inner (frameworks-dom-xhr-event-breakpoints.js:73)
    1) addFewBlackboxedListenersAndClick (frameworks-dom-xhr-event-breakpoints.js:78)
    2)  (:1)
    [setTimeout]
    0)  (:1)

Executing StepOut...
Call stack:
    0) addFewBlackboxedListenersAndClick (frameworks-dom-xhr-event-breakpoints.js:78)
    1)  (:1)
    [setTimeout]
    0)  (:1)

Executing Resume...
